requirements.txt
eve
gunicorn
flask-cors
bcrypt
py-bcrypt
pyjwt
bokeh
pyproj
requests